# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Dschang's Signal API
  description: REST API backend for the Dschang's Signal citizen issue reporting platform.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      operationId: getAllUsers
      tags: [ User ]
      summary: Get paginated list of users
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
      responses:
        '200':
          description: Paginated list of users
          headers:
            X-Total-Count:
              $ref: '#/components/headers/X-Total-Count'
            X-Page-Size:
              $ref: '#/components/headers/X-Page-Size'
            X-Page-Number:
              $ref: '#/components/headers/X-Page-Number'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      operationId: getUserById
      tags: [ User ]
      summary: Get a specific user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 4
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateUser
      tags: [ User ]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Cannot modify another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/role:
    patch:
      operationId: changeUserRole
      tags: [ User ]
      summary: Change a user's role (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/password:
    patch:
      operationId: updatePassword
      tags: [ User ]
      summary: Update user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Cannot modify another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [ User ]
      summary: Get the currently authenticated user
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Authenticated user found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/public:
    get:
      operationId: getPublicReports
      tags: [ Report ]
      summary: Get all validated reports
      security: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            default: created_at
      responses:
        '200':
          description: List of validated reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
  /reports/{id}:
    get:
      operationId: getReportById
      tags: [ Report ]
      summary: Get a specific report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Report found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteReport
      tags: [ Report ]
      summary: Delete a report
      description: Only the author of the report can delete it, and only if the report status is PENDING.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the report to delete
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Report deleted successfully
        '400':
          description: Report cannot be deleted because its status is not PENDING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - only the author can delete this report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports:
    post:
      operationId: createReport
      tags: [ Report ]
      summary: Create a new report
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{id}/moderation-status:
    patch:
      operationId: updateModerationStatus
      tags: [ Report ]
      summary: Update report moderation status (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModerationStatusRequest'
      responses:
        '200':
          description: Report status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{id}/report-status:
    patch:
      operationId: updateReportStatus
      tags: [ Report ]
      summary: Update report status (Admin only)
      description: Update the progression status (PENDING, IN_PROGRESS, RESOLVED) of a specific report.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusRequest'
      responses:
        '200':
          description: Report status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User does not have ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/admin:
    get:
      operationId: getAllReports
      tags: [ Report ]
      summary: Get all reports (Admin only)
      description: Retrieve all reports in the system with pagination and filtering.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: reportStatus
            example: createdAt,desc
        - name: moderationStatus
          in: query
          required: false
          schema:
            type: string
            enum: [ PENDING_REVIEW, ACCEPTED, REJECTED ]
        - name: reportStatus
          in: query
          required: false
          schema:
            type: string
            enum: [ PENDING, IN_PROGRESS, RESOLVED ]
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: fromDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful operation
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Page-Number:
              schema:
                type: integer
            X-Page-Size:
              schema:
                type: integer
            X-Total-Pages:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized - Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User does not have ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-Total-Count:
      schema:
        type: integer
      description: Total amount of elements for this query
    X-Page-Number:
      schema:
        type: integer
      description: Page number starting from 0
    X-Page-Size:
      schema:
        type: integer
      description: Page size
  parameters:
    Page:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
    Sort:
      name: sort
      in: query
      description: 'Sorting criteria (example: fullName)'
      required: false
      schema:
        type: string
        default: 'fullName'
  schemas:
...
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
          example: Duval Moi
        email:
          type: string
          format: email
          example: moi@u2g.com
    UpdatePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          example: OldP@ssword1
        newPassword:
          type: string
          format: password
          example: NewP@ssword1
    ChangeRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [ ADMIN, CITIZEN ]
          example: CITIZEN
    ReportRequest:
      type: object
      required:
        - title
        - description
        - locationText
      properties:
        title:
          type: string
          example: "Pothole on Kennedy Avenue"
          minLength: 3
          maxLength: 150
        description:
          type: string
          example: "Large dangerous pothole on the roadway"
          minLength: 10
        locationText:
          type: string
          example: "Kennedy Avenue, in front of the university"
        categoryIds:
          type: array
          items:
            type: integer
            format: int64
          example: [ 1, 2 ]
    UpdateModerationStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ PENDING_REVIEW, ACCEPTED, REJECTED ]
          example: ACCEPTED
        rejectionReason:
          type: string
          example: "Reporting outside area of jurisdiction"
    UpdateReportStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ PENDING, IN_PROGRESS, RESOLVED ]
          example: IN_PROGRESS
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: moi@u2g.com
        fullName:
          type: string
          example: Duval Moi
        role:
          type: string
          example: ADMIN
        createdAt:
          type: string
          format: date-time
          example: 2026-03-09T10:15:30Z
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: 2026-03-09T12:20:00Z
        isActive:
          type: boolean
          example: true
    Report:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Pothole on Kennedy Avenue"
        description:
          type: string
          example: "Large dangerous pothole on the roadway"
        locationText:
          type: string
          example: "Kennedy Avenue, in front of the university"
        moderationStatus:
          type: string
          enum: [ PENDING_REVIEW, ACCEPTED, REJECTED ]
          example: "PENDING_REVIEW"
        reportStatus:
          type: string
          enum: [ PENDING, IN_PROGRESS, RESOLVED ]
          example: "PENDING"
        rejectionReason:
          type: string
          example: "Area not covered by municipality"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-03-10T14:30:00Z"
        reviewedAt:
          type: string
          format: date-time
          example: "2026-03-11T09:15:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-10T14:30:00Z"
        createdBy:
          $ref: "#/components/schemas/User"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryResponse"
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Road"
          maxLength: 255
        color:
          type: string
          example: "#FF5733"
          pattern: '^#[0-9A-Fa-f]{6}$'
          maxLength: 10
        createdAt:
          type: string
          format: date-time
          example: "2026-03-10T10:30:00Z"
        createdBy:
          $ref: '#/components/schemas/User'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string